fix datatype

+++ griddata/grid.py
@@ -135,7 +135,7 @@ class Grid(object):
         grid.center = center
 
         # prepare for large array for storing the rotated map and prevent cropping upon rotation
-        big_shape = np.ones(self.ndim) * np.max([self.shape, shape])
+        big_shape = np.ones(self.ndim, dtype=np.integer) * np.max([self.shape, shape])
         ndelements = np.zeros(big_shape)
         offset = [int((big_shape[0] - self.shape[0]) / 2.0),
                   int((big_shape[1] - self.shape[1]) / 2.0),
